requirements.txt
pytest
requests
jsonschema
pydantic
pytest-pyreport
sqlalchemy
psycopg2-binary
python-dotenv
